pyproject.toml
[tool.poetry]
name = "ennuf"
version = "0.1.1"
description = "Easy Neural Networks in Fortran, a package for translating ML models"
authors = ["Helena Reid <helena.reid@metoffice.gov.uk>"]
readme = "README.md"
packages = [
    { include = "ennuf", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
tensorflow-cpu = "^2.12"
numpy = "^1.23"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
whl2conda = "^23.8.8"
black = "^23.9.1"
pylint = "^2.17.6"
mkdocs-material = "^9.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.pylint.format]
# Maximum number of characters on a single line, higher than default because default is extremely low.
max-line-length = 120

[tool.pylint.variables]
ignored-argument-names="arg|args|kwargs|_.*|^ignored_|^unused_"

[tool.pylint.messages_control]
disable = [
    # defalt values from pylint
    "raw-checker-failed", "bad-inline-option", "locally-disabled", "file-ignored", "suppressed-message",
    "useless-suppression", "deprecated-pragma", "use-symbolic-message-instead",
    # project specific values
    "missing-module-docstring"
]